backend/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.23
asyncpg>=0.29.0
psycopg>=3.1.12

# Document Processing
unstructured>=0.10.8
python-magic>=0.4.27
chromadb>=0.4.14

# Machine Learning & AI
openai>=1.3.0
langchain>=0.0.325
pydantic>=2.4.2
pydantic-settings>=2.0.3

# Utilities
python-jose>=3.3.0
passlib>=1.7.4
python-dotenv>=1.0.0
aiofiles>=23.2.1

# Testing (Optional)
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.1
